+++ src/main/kotlin/com/jlima/bookstoremanager/core/Author.kt
@@ -1,4 +1,4 @@
-package com.jlima.bookstoremanager.core.domain
+package com.jlima.bookstoremanager.core
 
 import java.util.Date
