+++ src/main/kotlin/com/jlima/bookstoremanager/core/Gender.kt
@@ -1,4 +1,4 @@
-package com.jlima.bookstoremanager.core.domain
+package com.jlima.bookstoremanager.core
 
 enum class Gender {
     MALE, FEMALE, OTHER
