+++ src/main/kotlin/com/jlima/bookstoremanager/service/AuthorService.kt
@@ -1,9 +1,9 @@
 package com.jlima.bookstoremanager.service
 
-import com.jlima.bookstoremanager.dto.domain.AuthorDTO
+import com.jlima.bookstoremanager.dto.AuthorDTO
 import com.jlima.bookstoremanager.providers.entity.domain.toDTO
 import com.jlima.bookstoremanager.providers.entity.domain.toEntity
-import com.jlima.bookstoremanager.providers.repository.domain.AuthorRepository
+import com.jlima.bookstoremanager.providers.repository.AuthorRepository
 import org.springframework.stereotype.Service
 import java.util.UUID
